requirements.txt
# Nowva AI Smart Squat Rack - Complete Requirements
# Python 3.11.2 recommended

# =============================================================================
# VOICE AGENT - GPT-4 Realtime API
# =============================================================================
livekit>=0.17.0
livekit-agents>=0.10.0
livekit-plugins-openai>=0.7.0          # GPT-4 Realtime API (handles STT+LLM+TTS)

# Note: These are no longer needed with Realtime API:
# - livekit-plugins-deepgram (STT handled by Realtime)
# - livekit-plugins-silero (VAD built into Realtime)
# - pyaudio (audio routing handled by LiveKit)

# =============================================================================
# DATABASE & SESSION MANAGEMENT
# =============================================================================
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0                 # PostgreSQL adapter
pydantic>=2.0.0                        # Data validation
cryptography>=41.0.0                   # Session encryption

# =============================================================================
# POSE ESTIMATION & BIOMECHANICS
# =============================================================================
# Computer Vision
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# MMPose (install via: mim install mmengine "mmcv>=2.0.0" "mmdet>=3.0.0" "mmpose>=1.0.0")
openmim>=0.3.0

# Numerical Computing
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# =============================================================================
# FASTAPI BACKEND (for program generation)
# =============================================================================
fastapi>=0.109.0                       # Web framework
uvicorn[standard]>=0.27.0              # ASGI server
httpx>=0.26.0                          # Async HTTP client (for voice agent → FastAPI)
openai>=1.0.0                          # OpenAI API client
tiktoken>=0.5.0                        # Token counting for OpenAI models

# =============================================================================
# CONTEXTUAL RAG SYSTEM
# =============================================================================
chromadb>=0.4.24                       # Vector database
voyageai>=0.2.0                        # Voyage AI embeddings
cohere>=5.0.0                          # Cohere reranking
rank-bm25>=0.2.2                       # BM25 lexical search
anthropic>=0.18.0                      # Claude API for contextualization

# =============================================================================
# UTILITIES
# =============================================================================
python-dotenv>=1.0.0                   # Environment variables
aiohttp>=3.9.0                         # Async HTTP
tqdm>=4.65.0                           # Progress bars
websockets>=12.0                       # WebSocket support

# =============================================================================
# OPTIONAL - DEVELOPMENT & TESTING
# =============================================================================
# jupyter>=1.0.0                       # Jupyter notebooks
# ipykernel>=6.23.0                    # Jupyter kernel
# pytest>=7.3.0                        # Testing
# black>=23.0.0                        # Code formatting
# flake8>=6.0.0                        # Linting

# =============================================================================
# OPTIONAL - WEB DASHBOARD (if needed)
# =============================================================================
# dash>=2.14.0                         # Interactive web dashboard
# plotly>=5.17.0                       # Plotly for visualizations

# =============================================================================
# INSTALLATION NOTES
# =============================================================================
# 1. Install base requirements:
#    pip install -r requirements.txt
#
# 2. Install MMPose (required for pose estimation):
#    mim install mmengine "mmcv>=2.0.0" "mmdet>=3.0.0" "mmpose>=1.0.0"
#
# 3. For M1/M2/M3 Macs, PyTorch may need special installation:
#    pip install torch torchvision torchaudio
#
# 4. If you need OpenSim (for advanced biomechanics - Week 6-7):
#    conda install -c opensim-org opensim
#    (Note: Not required for basic functionality)
#
# 5. For audio device testing (optional):
#    brew install portaudio  # macOS
#    pip install sounddevice
